+++ paywall/src/unlock.js/PostMessageEmitters/AccountsIframeMessageEmitter.ts
@@ -11,44 +11,26 @@ import {
   PostOfficeWindow,
   OriginWindow,
 } from '../../windowTypes'
-import {
-  makeIframe,
-  addIframeToDocument,
-  hideIframe,
-  showIframe,
-} from '../iframeManager'
+import { makeIframe, addIframeToDocument } from '../iframeManager'
 import {
   UserAccountsIframeEventEmitter,
   UserAccountsIframeEvents,
 } from './EventEmitterTypes'
+import { waitFor } from '../../utils/promises'
 
 // eslint is too stupid to parse this if the extends is in the class declaration below, so we extract it
 class FancyEmitter extends (EventEmitter as {
   new (): UserAccountsIframeEventEmitter
 }) {}
 
-/**
- * This class loads user accounts on-demand
- *
- * As such, it is slightly different from the other emitters.
- * This one does not create the iframe in the constructor, but instead
- * makes a fake one so that we can interact with it as if it were there,
- * but nothing happens unless we actually have an explicit need for user
- * accounts.
- *
- * The Wallet class handles initialization of this, and the
- * IframeHandler class handles setting up cross-iframe communication
- * in the IframeHandler.setupAccountUIHandler() method, which is also
- * called by the Wallet class in Wallet.setupWallet()
- */
-export default class AccountsIframeMessageEmitter extends FancyEmitter {
+export default class UserAccountsIframeMessageEmitter extends FancyEmitter {
   private _addHandler?: (
     type: keyof UserAccountsIframeEvents,
     listener: PostMessageListener
-  ) => void
+  ) => void = () => {}
   private window: IframeManagingWindow & PostOfficeWindow & OriginWindow
 
-  private _postMessage?: PostMessageResponder<PostMessages>
+  private _postMessage?: PostMessageResponder<PostMessages> = () => {}
   public iframe: IframeType
 
   constructor(
@@ -63,23 +45,18 @@ export default class AccountsIframeMessageEmitter extends FancyEmitter {
         postMessage: () => {},
       },
       className: '',
-      name: 'unlock accounts',
       src: userIframeUrl,
       setAttribute: () => {},
     }
   }
 
-  /**
-   * This is called by the Wallet.setupWallet() method if user accounts
-   * are needed
-   */
   createIframe() {
     const url = new URL(this.iframe.src)
-    this.iframe = makeIframe(this.window, this.iframe.src, 'unlock accounts')
+    this.iframe = makeIframe(this.window, this.iframe.src)
     addIframeToDocument(this.window, this.iframe)
 
     const { postMessage, addHandler } = mainWindowPostOffice(
-      this.window,
+      window,
       this.iframe,
       url.origin,
       'main window',
@@ -90,27 +67,24 @@ export default class AccountsIframeMessageEmitter extends FancyEmitter {
     this.setupListeners()
   }
 
-  /**
-   * This method is used by the MainWindowHandler to show the account iframe when needed.
-   * It also hides the checkout iframe if it is visible
-   */
   showIframe() {
     // note: if we are using a dummy iframe this will not display anything
-    showIframe(this.window, this.iframe)
+    this.iframe.className = 'unlock start show'
   }
 
   hideIframe() {
-    // note: if we are using a dummy iframe this will not hide anything
-    hideIframe(this.window, this.iframe)
+    this.iframe.className = 'unlock start'
   }
 
-  private setupListeners() {
+  async setupListeners() {
+    await waitFor(() => this._addHandler)
+    if (!this._addHandler) return
     this.addHandler(PostMessages.READY, () => this.emit(PostMessages.READY))
     this.addHandler(PostMessages.UPDATE_ACCOUNT, account =>
       this.emit(PostMessages.UPDATE_ACCOUNT, account)
     )
-    this.addHandler(PostMessages.UPDATE_NETWORK, network =>
-      this.emit(PostMessages.UPDATE_NETWORK, network)
+    this.addHandler(PostMessages.UPDATE_NETWORK, account =>
+      this.emit(PostMessages.UPDATE_NETWORK, account)
     )
     this.addHandler(PostMessages.INITIATED_TRANSACTION, () =>
       this.emit(PostMessages.INITIATED_TRANSACTION)
@@ -123,9 +97,6 @@ export default class AccountsIframeMessageEmitter extends FancyEmitter {
     )
   }
 
-  /**
-   * This is a proxy that ignores requests if the account iframe is not active yet
-   */
   async addHandler(
     type: keyof UserAccountsIframeEvents,
     listener: PostMessageListener
@@ -134,9 +105,6 @@ export default class AccountsIframeMessageEmitter extends FancyEmitter {
     this._addHandler(type, listener)
   }
 
-  /**
-   * This is a proxy that ignores requests if the account iframe is not active yet
-   */
   async postMessage<T extends MessageTypes = MessageTypes>(
     type: T,
     payload: ExtractPayload<T>
