+++ paywall/src/unlock.js/PostMessageEmitters/CheckoutIframeMessageEmitter.ts
@@ -11,12 +11,7 @@ import {
   PostOfficeWindow,
   OriginWindow,
 } from '../../windowTypes'
-import {
-  makeIframe,
-  addIframeToDocument,
-  showIframe,
-  hideIframe,
-} from '../iframeManager'
+import { makeIframe, addIframeToDocument } from '../iframeManager'
 import {
   CheckoutIframeEventEmitter,
   CheckoutIframeEvents,
@@ -30,15 +25,7 @@ class FancyEmitter extends (EventEmitter as {
   new (): CheckoutIframeEventEmitter
 }) {}
 
-/**
- * This is an abstraction layer around the post office for the checkout iframe
- *
- * It is used both to listen for incoming messages, and to send outgoing messages.
- * In addition, it handles showing and hiding the iframe. Those pieces are handled
- * in the MainWindowHandler for showing the checkout and user accounts iframes
- */
 export default class CheckoutIframeMessageEmitter extends FancyEmitter {
-  private window: IframeManagingWindow & PostOfficeWindow & OriginWindow
   public readonly addHandler: (
     type: keyof CheckoutIframeEvents,
     listener: PostMessageListener
@@ -53,8 +40,7 @@ export default class CheckoutIframeMessageEmitter extends FancyEmitter {
   ) {
     super()
 
-    this.window = window
-    this.iframe = makeIframe(window, checkoutIframeUrl, 'unlock checkout')
+    this.iframe = makeIframe(window, checkoutIframeUrl)
     addIframeToDocument(window, this.iframe)
 
     const { postMessage, addHandler } = mainWindowPostOffice(
@@ -69,11 +55,11 @@ export default class CheckoutIframeMessageEmitter extends FancyEmitter {
   }
 
   showIframe() {
-    showIframe(this.window, this.iframe)
+    this.iframe.className = 'unlock start show'
   }
 
   hideIframe() {
-    hideIframe(this.window, this.iframe)
+    this.iframe.className = 'unlock start'
   }
 
   async setupListeners() {
