+++ paywall/src/unlock.js/PostMessageEmitters/DataIframeMessageEmitter.ts
@@ -23,18 +23,24 @@ interface UnvalidatedPayload {
   params?: any
 }
 
+export function validateMethodCall(payload: UnvalidatedPayload) {
+  if (!payload || typeof payload !== 'object') return
+  if (!payload.method || typeof payload.method !== 'string') {
+    return false
+  }
+  if (!payload.params || !Array.isArray(payload.params)) {
+    return false
+  }
+  if (typeof payload.id !== 'number' || Math.round(payload.id) !== payload.id) {
+    return false
+  }
+  return true
+}
+
 class FancyEmitter extends (EventEmitter as {
   new (): DataIframeEventEmitter
 }) {}
 
-/**
- * This is an abstraction layer around the post office for the data iframe
- *
- * It is used both to listen for incoming messages, and to send outgoing messages.
- * For most, it simply emits what it receives, but in some cases it performs
- * additional logic, such as validating web3 method calls, or not forwarding
- * empty locks
- */
 export default class DataIframeMessageEmitter extends FancyEmitter {
   public readonly addHandler: (
     type: keyof DataIframeEvents,
@@ -50,7 +56,7 @@ export default class DataIframeMessageEmitter extends FancyEmitter {
   ) {
     super()
 
-    this.iframe = makeIframe(window, dataIframeUrl, 'unlock data')
+    this.iframe = makeIframe(window, dataIframeUrl)
     addIframeToDocument(window, this.iframe)
 
     const { postMessage, addHandler } = mainWindowPostOffice(
@@ -62,26 +68,6 @@ export default class DataIframeMessageEmitter extends FancyEmitter {
     )
     this.postMessage = postMessage
     this.addHandler = addHandler
-  }
-
-  /**
-   * Validate a payload sent from the data iframe to ensure it is in fact a web3 method call
-   */
-  validateWeb3MethodCall(payload: UnvalidatedPayload) {
-    if (!payload || typeof payload !== 'object') return false
-    if (!payload.method || typeof payload.method !== 'string') {
-      return false
-    }
-    if (!payload.params || !Array.isArray(payload.params)) {
-      return false
-    }
-    if (
-      typeof payload.id !== 'number' ||
-      Math.round(payload.id) !== payload.id
-    ) {
-      return false
-    }
-    return true
   }
 
   public setupListeners() {
@@ -101,7 +87,7 @@ export default class DataIframeMessageEmitter extends FancyEmitter {
     )
     this.addHandler(PostMessages.WEB3, payload => {
       // don't pass on anything that is not a valid web3 request
-      if (!this.validateWeb3MethodCall(payload)) return
+      if (!validateMethodCall(payload)) return
       this.emit(PostMessages.WEB3, payload)
     })
     this.addHandler(PostMessages.UPDATE_ACCOUNT, account =>
