+++ paywall/src/unlock.js/PostMessageEmitters/EventEmitterTypes.ts
@@ -3,15 +3,6 @@ import StrictEventEmitter from 'strict-event-emitter-types'
 import { PostMessages, ExtractPayload } from '../../messageTypes'
 import { web3MethodCall } from '../../windowTypes'
 
-/**
- * These are the event definitions for the data iframe. They correspond directly with the post message events
- * These events are the ones received from the data iframe
- *
- * for new events, they should have 1 of 2 forms:
- *
- * for events with no payload, () => void
- * for all others, (payload: ExtractPayload<PostMessages.***>) => void
- */
 export interface DataIframeEvents {
   [PostMessages.READY]: () => void
   [PostMessages.READY_WEB3]: () => void
@@ -35,19 +26,9 @@ export interface DataIframeEvents {
   [PostMessages.UPDATE_NETWORK]: (
     update: ExtractPayload<PostMessages.UPDATE_NETWORK>
   ) => void
-  // this has to be more specific because WEB3 is overloaded
   [PostMessages.WEB3]: (request: web3MethodCall) => void
 }
 
-/**
- * These are the event definitions for the checkout iframe. They correspond directly with the post message events
- * These events are the ones received from the checkout iframe
- *
- * for new events, they should have 1 of 2 forms:
- *
- * for events with no payload, () => void
- * for all others, (payload: ExtractPayload<PostMessages.***>) => void
- */
 export interface CheckoutIframeEvents {
   [PostMessages.READY]: () => void
   [PostMessages.DISMISS_CHECKOUT]: () => void
@@ -56,15 +37,6 @@ export interface CheckoutIframeEvents {
   ) => void
 }
 
-/**
- * These are the event definitions for the user accounts iframe. They correspond directly with the post message events
- * These events are the ones received from the user accounts iframe
- *
- * for new events, they should have 1 of 2 forms:
- *
- * for events with no payload, () => void
- * for all others, (payload: ExtractPayload<PostMessages.***>) => void
- */
 export interface UserAccountsIframeEvents {
   [PostMessages.READY]: () => void
   [PostMessages.UPDATE_ACCOUNT]: (
